Updated nuclear density distribution models for some light nuclei. Added three parameter Fermi model and sum-of-Gaussian model. Also, added the form factor for sum-of-Gaussian model.
I updated nuclear density distribution models for some light nuclei. I worked on H-2, He-4, Be-10, B-11, C-12, N-14, O-16. I added three parameter Fermi model for  N-14 & O-16 and the sum-of-Gaussian model for He-4 & C-12 and updated the default radius to the measured rms radius for the remaining nuclei following the Gaussian model (H-2, Be-10, and B-11 ). Also, I added the form factor for the sum-of-Gaussian model.